Refactored whole coreblocks code.
