+++ coreblocks/priv/csr/csr_register.py
@@ -8,16 +8,16 @@ from transactron.utils import assign, bits_from_int
 from coreblocks.params.genparams import GenParams
 from transactron.utils.dependencies import DependencyManager, ListKey
 from coreblocks.params.fu_params import BlockComponentParams
-from coreblocks.params.layouts import FetchLayouts, FuncUnitLayouts, CSRLayouts
-from coreblocks.params.isa import Funct3, ExceptionCause
-from coreblocks.params.keys import (
+from coreblocks.interface.layouts import FetchLayouts, FuncUnitLayouts, CSRLayouts
+from coreblocks.frontend.decoder.isa import Funct3, ExceptionCause
+from coreblocks.interface.keys import (
     AsyncInterruptInsertSignalKey,
     FetchResumeKey,
     ExceptionReportKey,
     InstructionPrecommitKey,
 )
-from coreblocks.params.optypes import OpType
-from coreblocks.utils.protocols import FuncBlock
+from coreblocks.frontend.decoder.optypes import OpType
+from coreblocks.func_blocks.interface.func_protocols import FuncBlock
 from transactron.utils.transactron_helpers import from_method_layout
